refactor#61 : User domain 분리
User Domain을 분리했다
