+++ src/user/user.controller.ts
@@ -1,5 +1,7 @@
 import express, { Request, Response } from 'express';
-import { AbstractUserController, AbstractUserService } from './user';
+
+import { AbstractUserController } from './type/userController';
+import { AbstractUserService } from './type/userService';
 
 export default class UserController implements AbstractUserController {
 	private static instance: AbstractUserController;
@@ -27,7 +29,6 @@ export default class UserController implements AbstractUserController {
 		UserController.router.patch('/:id', this.updateById);
 		UserController.router.delete('/:id/thumbnail', this.deleteThumbnail);
 		UserController.router.patch('/:id/thumbnail', this.updateThumbnail);
-
 		app.use(UserController.PATH, UserController.router);
 	}
 
@@ -53,8 +54,8 @@ export default class UserController implements AbstractUserController {
 	}
 
 	async updateById(req: Request, res: Response): Promise<void> {
-		const id = Number(req.params.id);
 		try {
+			const id = Number(req.params.id);
 			await UserController.userService.updateById(id, req.body);
 			res.status(201).send();
 		} catch (error) {
@@ -63,9 +64,13 @@ export default class UserController implements AbstractUserController {
 	}
 
 	async getById(req: Request, res: Response): Promise<void> {
-		const id = Number(req.params.id);
-		const findUser = await UserController.userService.findById(id);
-		res.send(findUser);
+		try {
+			const id = Number(req.params.id);
+			const findUser = await UserController.userService.findById(id);
+			res.status(200).send(findUser);
+		} catch (error) {
+			res.status(400).send();
+		}
 	}
 
 	async getByNickname(req: Request, res: Response): Promise<void> {
